+++ framework-utilities/src/main/java/uk/gov/justice/services/framework/utilities/exceptions/StackTraceProvider.java
@@ -1,0 +1,10 @@
+package uk.gov.justice.services.framework.utilities.exceptions;
+
+import org.apache.commons.lang3.exception.ExceptionUtils;
+
+public class StackTraceProvider {
+
+    public String getStackTrace(final Throwable throwable) {
+        return ExceptionUtils.getStackTrace(throwable);
+    }
+}