+++ jmx/jmx-api/src/main/java/uk/gov/justice/services/jmx/api/UnsupportedSystemCommandException.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.jmx.api;
+
+public class UnsupportedSystemCommandException extends RuntimeException {
+
+    public UnsupportedSystemCommandException(final String message) {
+        super(message);
+    }
+}